Update updateTags function by using new runtime function (#67)
Issue #, if available:
New method of comparing tags was added in runtime. Using this new method to reduce amount of custom logic code of updateTags.

Description of changes:
Use new method in runtime to replace custom logic of comparing two tags.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.