Delete `tests/e2e` directory (#70)
Description of changes:
The `tests/e2e` directory is an incorrectly named version of the `test/e2e` directory. It should be removed so as to not cause confusion for contributors who might not know where tests should be added.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.