Return an error message if snapshot is not ready for deleting (#77)
Issue #, if available:
If `snapshot` is not ready for deleting and is called by `sdkDelete`, it will have an error `InvalidSnapshotStateFault`.

Description of changes:
Check whether `snapshot` is in the `status` to be deleted. If not, then output an error message.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.